# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API La Remise",
    "version": "1.0.0",
    "description": "API de gestion de la ressourcerie La Remise"
  },
  "paths": {
...
    "/api/depots": {
      "post": {
        "summary": "Enregistre un dépôt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["personne_id", "date_depot", "type"],
                "properties": {
                  "personne_id": { "type": "integer" },
                  "date_depot": { "type": "string", "format": "date" },
                  "type": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dépôt créé" },
          "400": { "description": "Champs obligatoires manquants" }
        }
      }
    },
    "/api/depots/{id}": {
      "get": {
        "summary": "Récupère un dépôt par son id (avec objets et donatrice)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Id du dépôt"
          }
        ],
        "responses": {
          "200": { "description": "Le dépôt demandé" },
          "400": { "description": "Id invalide" },
          "404": { "description": "Dépôt introuvable" }
        }
      }
    },
    "/api/depots/{id}/objets": {
      "post": {
        "summary": "Ajoute un objet au dépôt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Id du dépôt"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["libelle", "poids_kg", "etat_arrivee", "categorie_id"],
                "properties": {
                  "libelle": { "type": "string" },
                  "poids_kg": { "type": "number" },
                  "etat_arrivee": { "type": "string" },
                  "categorie_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Objet ajouté au dépôt" },
          "400": { "description": "Champs obligatoires manquants" }
        }
      }
    },
...
  }
}